Add condition to prevent running frontend users migrations if rainlab user plugin is not installed

+++ updates/update_frontend_users_table.php
@@ -7,14 +7,16 @@ class UpdateFrontEndUsersTable extends Migration
 {
     public function up()
     {
-        Schema::table('users', function ($table) {
-            $table->timestamp('last_activity')->nullable();
-        });
+        if (class_exists('Rainlab\User\Models\User')) {
+            Schema::table('users', function ($table) {
+                $table->timestamp('last_activity')->nullable();
+            });
+        }
     }
 
     public function down()
     {
-        if (Schema::hasColumn('users', 'last_activity')) {
+        if (class_exists('Rainlab\User\Models\User') && Schema::hasColumn('users', 'last_activity')) {
             Schema::table('users', function ($table) {
                 $table->dropColumn('last_activity');
             });
